Bugs bugs bugs
Tiles are loading but aren't rendering. If you go into PlayerEntity and
print out (standingAt) it will say the tile is tile.grass. On the image
this is correct, its just not rendering.

method loadTiles in Level class at line 83 is where the tiles are added
based on levelColour. More explanation at the method.